getElementByClassName finished
Document.body may be an object but it has too much stuff to put into the
recursion. It's got other things chocked in it. The pseudocode was if this
is a collection...But the code was, does this have childnodes.
If this has child nodes, bring it up with the childNodes method because
body is an object.